New '/assemblylist' switch to emit and show assembly list instead of full solution tree (move bits of rendering/sorting out of 'SolutionGenerator' class, parametrize 'FinalizeProjects()')
